Move utf-8 decoding step inside cython extension
The cython setters take python strings, but the getters return
python bytes, so this makes the types match up. I went with regular
python strings for the interface since they are more common at the
python level, so the encoding/decoding is now an implementation detail
of the cython extension.